requirements.txt
openai>=1.0.0
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.16
langchain-core==0.2.38
duckduckgo-search>=4.0.0
requests>=2.31.0
httpx>=0.24.0
python-dotenv>=1.0.0
geopy>=2.4.0
qrcode[pil]>=7.4.2
Pillow>=10.0.0
pyTelegramBotAPI>=4.14.0

